Monkeypatch core cron provider with custom facts
This is the most heinous, abusive and evil thing I have ever done to Puppet.
But, it allows the parselet provider to take over without interference from
the core cron provider.

The monkey patch also fixes an issue with the purging method of the Cron Type.
This fix should be submitted upstream.