Add Puppet unit testing framework
This commit adds all the libraries needed to unit test Puppet modules:

  - rspec
  - mocha
  - puppetlabs_spec_helper
  - puppet
  - rspec-puppet

Really, only rspec and mocha are required at this point since I'm only testing
the parser and the parse is independent of any Puppet code. However, I want to
test the parser on crontabs stored as fixtures, so I'm bringing
`puppetlabs_spec_helper` in early and this necessitates bringing in `puppet`
and `rspec-puppet`.